☁️🚀 ﻿ノ( ゜-゜ノ) Initial commit
This cocoapods currently has 2 sample projects, one in Objective-C (/Dev) and one in Swift (/DevSwift). Eventually we will drop to one sample project + test (probably Swift), but the cocoapod itself is in Objective-C so it can be used more easily from iOS 7 codebases (which cannot load frameworks, a requirement of swift pods).