Hide UIView+LKAdditions behind internal flag
If this flag is available, then LaunchKit classes will reveal the UIView+LKAdditions properties (which "infect" all UIViews). The reason for hiding this behind LK_IB_DESIGNABLE is because if this pod is being used in a static library, it will show an error in Xcode (cannot you use IBInspectable/Designable in code that ends up in a static lib).

The error doesn't prevent a build, though. However it still looks bad so let's just make this a bit nicer looking to new developers who would frown at our SDK generating an error in their code.