More direct 'commit' approach to extracting analytics
This is so analytics can know that something is being 'committed', and it can flush its tap tracking and returning. In the past, the method was just 'trackableProperties' which implied "just what you have now", and weirdly LaunchKit instance would be in charge of clearing the properties from analytics, which seems like an implementation detail. So, this is more direct and analytics can be a bit smarter about what its returning.