Send up rotation-aware window info.
Previously we used to ONLY send up portrait-orientation windows, because our tap tracking would send up a 'l' or 'p' per tap, to hint at the transform to apply to the canonical window size to generate the canvas upon which to render the tap.

However, starting with iOS 9 and iPads, windows are no longer guaranteed to be full-screen. They can change size arbitrarily. So the idea of a canonical window size is no longer appropriate.