Create initial Carthage workspace
NOTE: The name of the workspace is alphabetically important. The way Carthage appears to search a repo for workspaces and projects is it that it finds the first project and first workspace (if available), alphabetically.

If a workspace is found, it prefers the first project in that. So naming the workspace "BuildForCarthage" ensures it has a good chance of being found first, compared to "Dev" and "DevSwift" workspaces.