Always reset the tracking timer, when tracking
This will reduce a number of superfluous tracking calls. For example: If tracking something due to user action, like changing resigning active and coming back, and then the tracking timer unrelatedly fires immediately after. It makes sense to push back the tracking timer each time we track, so it stays on the schedule based on the last time we tracked.