Add shadow support to Whats New (more in comments)
So the implementation is a bit goofy; because shadows cannot co-exist with rounded corners in the same view, I had to make What's New existed in a whole new containing subview, so the hierarchy is: LKViewController > shadowView > cardView.

Trying to do this in a generic way is a bit clumsy, but for now LKViewController has an optional outlet 'cardView' that can be set in IB, and if it's not set, it'll just point to the view controller's main view.

In card presentation controller, I am manually changing clipsToBounds of the main view, which is sort of odd, but for some reason iOS changes clipsToBounds to YES, even if I set it to NO in IB.