Add block-based refresh handler to config
This refresh handler will fire whenever the config has changed. It will *ALSO* first the FIRST time the config receives a new parameters dictionary, whether or not it was the same as our cached dictionary. This gives app developers a nice, single block from which to handle first-time (and subsequent actual) updates from the server