Add unit test to verify same session id after simultaneous "empty" track calls
This test is designed to test the sequential tracking call logic. Sending a track call without a `session_id` parameter will always generate a new session_id that is meant to be sent up in all subsequent track calls.

In the past, calling track one-after-another would result in 2 different session_ids, which created empty rows in the database, etc. Now we can ensure that the response of one track call is handled before another track call is sent.