Rewrite local bundle timestamp check _phew_
Before: We'd store a property for the last server time into NSUserDefaults, and check against that.

Problem with that is that it didn't tie the 'last server time' with what was in our cache (which could be accidentally or deliberately deleted/corrupted by iOS). There was the chance that we wouldn't have local bundles, yet assume that everything was fine because our NSUD-stored 'last server time' matched the remote server time.

After:
Pass along the server timestamp along our requests to retrieve the manifest and bundles. When the set is finished download, write out a text file with the server timestamp to the bundles download folder. Use this timestamp file as the "truth" of what is available locally.