Add special check for `LAUNCHKIT_MANUAL_IMPORT`
This is for folks that do not want to import LaunchKit as a framework or static library, but instead want to integrate a snapshot of LaunchKit code directly into their. While we don't recommend this (Carthage or Cocoapods is great! Try them!), this at least allows a developer to continue importing the LK codebase manually.