Add ability for LKImageView to constrain itself
This allows us to constrain an image view's height to it's width at the same aspect ratio as its image. So it essentially adds a constraint keeping the image view the same shape as its image.

This is going to very useful for handling images whose sizes don't fit perfectly within the layout bounds. By default UIImageViews will use the intrinsicContentSize at this point during auto-layout, but that can often be a very different size than the actual layout size.