Add support for onboarding UI rotation on iPads
This will remember the current "page" and reset to that page's content offset under the new orientation.

This is implemented in both iOS 7's willAnimateRotation... which works iOS 7.0+, however if the LK library or app is running on iOS 9 and up, the OS may ignore the iOS 7 animation calls and only call `viewWillTransitionToSize` instead, so implement the rotation stuff there too.