Fix bug causing some bundle loads to fail
When are asked to load a bundle name, we have an initial shortcut check. The idea is: "If we have the bundle manifest, we know what bundle we're supposed to have, and NOT have. If the manifest is available, then check if the bundle is NOT in our manifest, and early return with an error if not)"

The problem was that we were checking the bundle infos which represented bundles ALREADY DOWNLOADED, not bundle infos that are in our latest manifest (and possibly not downloaded). As such, sometimes the *first* time a bundle was requested, it might have failed it we were with a fresh cache, and actively downloading bundles.

This fix ensure that if the manifest is available, we check the MANIFEST bundle infos, to make a decision on whether we don't have that bundle (in order to fail early).

Phew.