Change the onboarding -> app transition logic
This change ensures that the app's vc becomes the rootViewController first, then we animate away the onboarding as if it's still the root view controller.

The old way was to animate the app vc in, pretending to be the rootViewController, then when the transition was over, to *actually* set the rootViewController then. The problem with this approach was that once in a while, UIKit would fire `viewDidAppear` calls to the app VC during that animation, but before that VC was set as the root VC. If that app VC was presenting UI during a `viewDidAppear` (e.g. presenting LK App release notes), then occasionally the parent view controller would get messed up.

This new approach ensures that the app VC immediately gets set as the rootVC, and the animation is changed such that the fromViewController is just removed from super view at the end.