Copy local bundles map as remote, if same time
We treat the remoteBundleMap as our "source of truth", for what is/should be available to the the system. But upon startup, if we get the same timestamp as before, our remoteBundleMap was never being set up. So in the first-time case, if we are supposed to be "up to date" with the server, then copy over what we have locally as our remote "source of truth" map, so that we can do lookups against it.