Add LKLabel component
This is needed for the app review card, currently, and could be useful to us in the future (to perhaps handle dynamic font sizing).

The reason it is needed for the app review card is that the message body label (multiline) will sometimes get the wrong "preferredMaxLayoutWidth" when it is set to 'automatic' on the storyboard. Somehow the calculating (within the LKSimpleStackView) becomes incorrect. The solution appears to be to have an explicit preferredMaxLayoutWidth, which is something we'd need for iOS 7 currently anyways).

This LKLabel has an IBInspectable property to optionally enable preferredMaxLayoutWidth being explicitly set during `layoutSubviews`