Don't show Release Notes to new (to-app) users
This change adds a check for the 2 cases we need to check to see if a user has use an older app version before:
1. LK `current_version_duration` is (meaningfully) shorter than `install_duration`

This means the user was an LK session in a previous version of the app, so they definitely have updated.

2. App install time (guessed inspecting creation date of Documents/ dir) is (meaningfully) longer than LK install time.

I made the check for about 1 day. If the app install time is less than a day, treat them as a new user. This is because the Documents folder gets created upon app download (I believe), and so the user may not run the app immediately.

This does mistake people that installed an app but took days to run it the first time (with LK) as people who have updated, and therefore those people *would* see Release Notes, even though they were essentially new users. I think this still reasonable, since the number of people that might do that are low (enough).