When changing app version or build, delete bundles
This fixes an issue with Release Notes where updating the app version number or build was still showing the previous version's bundle.

This is due to the fact that the bundles manager was only going off of the "lastBundleUpdatedTime" from the server. If the app had seen the latest bundlesupdatedTime while being version "A", and then later switched to version "B", it would think, "oh, I don't need to go and get a new manifest, since i'm 'up-to-date'", which was wrong (and needed, to download the release notes for version B. So, app version B was potentially showing cached release notes for version A.

With this fix, we track the last app version and build that we have seen in the bundles manager, and if the version is different, we purge the bundles cache, to ensure we always download the new manifest (and any bundles)