Fix bug where "forcePresentAppReleaseNotes" was not working
Thanks to @bryanbartow for the find! The issue was that, with the new, more robust changes to detect a new-to-app vs new-to-launchkit user, the "ignoreIfShownBefore" logic was being superceded by whether the user was a new-to-app user.

This change explicitly checks only the "whatsNewEnabled" config, if the request for release notes is a direct user request.