Improve measuring of LKVC's, and cache iOS 8 sizes
This fixes a bad bug on "measurable" LKVC's where on iOS 8, by simply measuring the UI (using `systemLayoutSizeFittingSize:withHorizontalPriority:verticalPriority`), was triggering another layout, and it was leading to an infinite loop on the main thread. Yikes.

This makes sure we're caching our size measurement by width, and returning the cached value on iOS 8.